Enable mapping TELL-Seq and DBS linked reads

+++ include/util.h
@@ -18,7 +18,7 @@
 
 #define IS_ACGT(c) ((c) == 'A' || (c) == 'C' || (c) == 'G' || (c) == 'T')
 
-typedef uint32_t bc_t;
+typedef uint64_t bc_t;
 typedef uint32_t chrom_t;
 
 void copy_until_space(char *dest, char **src);
